move layout folder to src
