+++ src/App.jsx
@@ -1,7 +1,7 @@
 import { useState } from 'react'
-import Header from './assets/Layout/Header/Header'
+import Header from './Layout/Header/Header'
 import { BrowserRouter, Routes, Route } from 'react-router-dom'
-import Home from './assets/pages/Principal'
+import Home from './pages/Principal'
 
 import './App.css'
