+++ src/Layout/Header/Header.jsx
@@ -1,6 +1,6 @@
 import React from "react";
-import LogoTitulo from "../../static/Imagens/LogoTitulo.png";
-import BotaoRosaPrincipal from "../../../components/PinkButton";
+import LogoTitulo from "../../Imagens/LogoTitulo.png";
+import BotaoRosaPrincipal from "../../components/PinkButton";
 import './HeaderCss.css'; // Importe o arquivo CSS
 
 const Header = (props) => {
